requirements.txt
# Database
sqlalchemy==2.0.23
pymysql==1.1.0

# Data & Environment
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0

# NLP
transformers==4.36.2
torch==2.1.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
shap==0.44.0

# Graph Neural Networks
torch-geometric==2.4.0
dgl==1.1.2

# MLOps
mlflow==2.9.2

# API
fastapi==0.104.1
uvicorn==0.24.0

# Dashboard
streamlit==1.29.0
plotly==5.18.0
folium==0.14.0
streamlit-folium==0.16.0

# Orchestration (installed separately)
# apache-airflow
# pyspark

colorlog